+++ src/totem.h
@@ -122,7 +122,7 @@ void	totem_action_next			(Totem *totem);
 void	totem_action_previous			(Totem *totem);
 void	totem_action_seek_time			(Totem *totem, gint64 sec);
 void	totem_action_seek_relative		(Totem *totem, int off_sec);
-void	totem_action_volume_relative		(Totem *totem, int off_pct);
+void	totem_action_volume_relative		(Totem *totem, double off_pct);
 gboolean totem_action_set_mrl			(Totem *totem,
 						 const char *mrl);
 void	totem_action_set_mrl_and_play		(Totem *totem,
